+++ frontend/package.json
@@ -10,7 +10,8 @@
     "react": "^18.1.0",
     "react-bootstrap": "^2.4.0",
     "react-dom": "^18.1.0",
-    "react-scripts": "^5.0.1",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^2.1.3",
     "uuid": "^8.3.2",
     "web-vitals": "^2.1.4"
   },
