+++ frontend/src/pages/labeling/Labeling.css
@@ -2,26 +2,18 @@
   text-align: center;
 }
 
-.NER_types_inputs {
-  height: 5vh;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
-}
-
 .ControlPanel {
-  height: 45vh;
+  height: 36vh;
   background-color: deepskyblue;
 }
 
 .Sentence {
   font-size: 1em;
   color: black;
-  height: 10vh;
-  width: 90%;
-  margin: 2%;
+  height: 8vh;
+  width: 96%;
+  margin-right: 2%;
+  margin-left: 2%;
   word-wrap: normal;
 }
 
@@ -47,7 +39,7 @@
 
 .LabelList {
   overflow-y: scroll;
-  height: 45vh;
+  height: 51vh;
 }
