+++ frontend/src/pages/learninge2e/LearningE2E.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export function Home() {
+    return (
+        <div>
+            <h2>Learning E2E</h2>
+            <ul>
+                <li>Use Admin page to upload NER types & upload/download data.</li>
+                <li>Use Labeling page for data labeling.</li>
+            </ul>
+        </div>
+    );
+}
